fix(cli): move cli into lib, allowing exports from lib/index.js
