+++ lib/cli/blueprint.js
@@ -5,7 +5,7 @@ import walk from 'walk-sync';
 import mkdirp from 'mkdirp';
 import rimraf from 'rimraf';
 import ui from './ui';
-import isDir from '../lib/utils/is-dir';
+import isDir from '../utils/is-dir';
 import template from 'lodash/template';
 
 /**
