+++ lib/cli/builder.js
@@ -6,8 +6,8 @@ import BabelTree from 'broccoli-babel-transpiler';
 import MergeTree from 'broccoli-merge-trees';
 import LintTree from './lint-tree';
 import PackageTree from './package-tree';
-import discoverAddons from '../lib/utils/discover-addons';
-import tryRequire from '../lib/utils/try-require';
+import discoverAddons from '../utils/discover-addons';
+import tryRequire from '../utils/try-require';
 
 
 export default class Builder {
