+++ lib/cli/command.js
