+++ lib/cli/lint-tree.js
