+++ lib/cli/package-tree.js
