+++ lib/cli/project.js
@@ -10,10 +10,10 @@ import { sync as copyDereferenceSync } from 'copy-dereference';
 import chalk from 'chalk';
 import ui from './ui';
 import Builder from './builder';
-import discoverAddons from '../lib/utils/discover-addons';
-import tryRequire from '../lib/utils/try-require';
-import isDir from '../lib/utils/is-dir';
-import eachDir from '../lib/utils/each-dir';
+import discoverAddons from '../utils/discover-addons';
+import tryRequire from '../utils/try-require';
+import isDir from '../utils/is-dir';
+import eachDir from '../utils/each-dir';
 
 export default class Project {
